+++ utils_function/utils.hpp
@@ -26,5 +26,5 @@ int		ft_atoi_chunk(const char *src);
 size_t	ft_find(const char *str, const char *find);
 void	*ft_memcpy(void *dst, const void *src, size_t n);
 char			*ft_strtrim(char *s1, char const *set);
-void 	*ft_memjoin(char *dst, char *src, int &dst_size, int src_size);
+void 	*ft_memjoin(char *dst, char *src, size_t &dst_size, size_t src_size);
 char *ft_erase(char *&dst, int dst_size, int len);
